N meetings in one room(DSA)
There is one meeting room in a firm. There are N meetings in the form of (start[i], end[i]) where start[i] is start time of meeting i and end[i] is finish time of meeting i.
find  the maximum number of meetings that can be accommodated in the meeting room when only one meeting can be held in the meeting room at a particular time.